split to modules, add ip - imgprocessing module
