+++ ip/__init__.py
@@ -1,0 +1,1 @@
+# ip - image processing module
